pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capa-dynamic-argument-matching-prototype"
version = "0.1.0"
description = "MVP prototype for structured call argument matching in dynamic analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
